Simplify/optimize reading/writing to/from a ByteBuffer.
