+++ clustering/server/src/main/java/org/wildfly/clustering/server/dispatcher/CommandDispatcherMarshaller.java
@@ -29,8 +29,8 @@ import java.nio.ByteBuffer;
 import org.jboss.marshalling.Marshaller;
 import org.jboss.marshalling.Marshalling;
 import org.wildfly.clustering.dispatcher.Command;
-import org.wildfly.clustering.marshalling.jboss.ByteBufferOutputStream;
 import org.wildfly.clustering.marshalling.jboss.MarshallingContext;
+import org.wildfly.clustering.marshalling.spi.ByteBufferOutputStream;
 import org.wildfly.clustering.marshalling.spi.IndexSerializer;
 import org.wildfly.clustering.marshalling.spi.MarshalledValueFactory;
